ISSUE #366 move param parsing to it's own file
